// src/app/model/mediopago.dto.ts
export interface Mediopagodto {
  id: number;
  nombre: string;
  descripcion: string;
  estado: string;
}
export interface MediopagoActualizadodto {
  id: number;
  nombre: string;
  descripcion: string;
  estado: string;
}
export interface MediopagoNuevodto {
  nombre: string;
  descripcion: string;
  estado: string;
}